alterando queries das demandas

+++ server/src/main/java/com/vincula/service/indicador/IndicadorResultadoService.java
@@ -26,7 +26,7 @@ public class IndicadorResultadoService {
         LocalDate inicio = filtro.getDataInicial();
         LocalDate fim = filtro.getDataFinal();
 
-        int totalFinalizadas =
+        Long totalFinalizadas =
                  demandaRepository.countDemandasFinalizadas(
                         unidadeResponsavelId,
                         unidadeSolicitanteId,
